Improve rendering of non lat-lon tiles
There have been a long standing bugs related to this..
Up until now the solution was to only use tiles that split along
lat-long lines and to also split the geometry (in roam) along the same
lines.

Hopefully this allows tiles to be rendered properly.

See examples/tex.c for details of how this rendering works.