Rewrite expose to preserve the depth buffer
Sorted objects are now drawn with depth testing enabled. Unsorted
objects are drawn without depth testing. This mostly affects
GRITS_LEVEL_WORD+n objects.

Objects at other levels such as BACKGROUND, OVERLAY, and HUD should
generally disable sorting so that they are always drawn to the screen.
For those levels, the depth-buffer is set to read-only. This preserves
the depth buffer so  that it can be used later on to get the world-level
depth at a given screen location.