Add grits_viewer_unproject function
This does the opposite of the grits_viewer_project function and can be
used to get the lat-lon-elev point corresponding to a point on the
screen, such as the pointer location. This currently has an undefined
output if the screen coordinates do not map to the location of an object
in the viewer.

This should not be used to determine if an objects has been selected or
clicked on. OpenGL picking should be used for that by connecting to one
of the GritsObject mouse event signals.