Add ortho flag to GritsMarker
Setting ortho to true causes the marker to be drawn orthographic and
flat against the screen. This is the default and behavior and is markers
have been drawn in the past.

Clearing the flag causes he marker to be drawn using the viewers
perspective projection but scaled to a fixed size.