Remove a lot of callback from GritsOpenGL
Many of the callbacks can be moved the expose function. Since they
mostly run in the main thread anyway it doesn't matter too much.

One distinct advantage of this is that the split_merge and expose
functions are interleaved, the end result of which is a much nicer
looking display.