Move low level tile loading to GritsTile
Move all the OpenGL stuff out of the plugins and into GritsTile, this
prevents needing to explicit idle callbacks in plugins, which is
extremely unstable when unloading a plugin.

Note: The tile code is _not_ thread safe and will be reworked in
subsequent changes. The plan is to use some sort of priority queue and
do all the updating and garbage collecting from the main thread directly
from the location-changed callback. The plugins can still use thread
pools for the actual downloading and loading.